Shape继承UVMap提供默认映射，UVMap不再必须；修复目录Bug，修复图像和棋盘格映射Bug；棋盘格n可以是小数；恢复以Debug模式编译

+++ Texture/GridTexture.h
@@ -9,10 +9,10 @@
 #include "Texture.h"
 
 class GridTexture: public Texture {
-    int n;
+    float n;
     Color a, b;
 public:
-    GridTexture(int n = 8, const Color &a = Color::zero, const Color &b = Color(1,1,1));
+    GridTexture(float n = 8, const Color &a = Color::zero, const Color &b = Color(1,1,1));
     Color operator()(Point const &uv) const override;
 };
